requirements.txt
aliyun-python-sdk-core==2.13.19
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.10.1
appdirs==1.4.4
apturl==0.5.2
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
command-not-found==0.3
crcmod==1.7
cryptography==2.9.2
cupshelpers==1.0
defer==1.0.6
defusedxml==0.7.0rc1
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
Django==2.2.13
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-request==1.5.5
filelock==3.0.12
gunicorn==20.0.4
httplib2==0.9.2
idna==2.9
importlib-metadata==1.7.0
importlib-resources==3.0.0
itsdangerous==1.1.0
jmespath==0.10.0
jsonfield==3.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
netifaces==0.10.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
oss2==2.12.1
pbr==5.4.5
pexpect==4.2.1
Pillow==7.1.2
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.7
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.13.2
six==1.15.0
social-auth-core==3.3.3
sqlparse==0.3.1
stevedore==3.0.0
swapper==1.1.2.post1
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.3
virtualenv==20.0.26
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.2
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
